+++ test/product.spec.js
@@ -1,11 +1,25 @@
 const { expect } = require("chai");
 
-const app = require("supertest")(require("../server/server"));
+
+const { syncAndSeed } = require('../server/db');
+
+const _app = require('../server/app')
+const app = require("supertest")(_app);
+
+const axios = require('axios');
+
 
 const {
     db,
     model: { Products },
 } = require("../server/db");
+
+
+describe('Testing', () => {
+    it('equals 2', ()=>{
+        expect(1 + 1).to.equal(2)
+    })
+});
 
 describe("Backend", () => {
     describe("Product Model", () => {
@@ -28,3 +42,23 @@ describe("Backend", () => {
         });
     });
 });
+
+
+describe('Routes', () => {
+    describe('GET /', () => {
+        beforeEach(async () => {
+            await db.sync({ force: true });
+            //await syncAndSeed()
+          });
+     
+        it('testing get route', async() => {
+            const response = await app.get('/api/products');
+            expect(response.status).to.equal(200);
+        })
+        it('compare axios and get', async() => {
+            const response = await app.get('/api/products');
+            const products = await Products.findAll();
+            expect(response.body.length).to.equal(products.length)
+        })
+    })
+})
